requirements.txt
requests
pandas
numpy
ccxt
python-dotenv
